+++ CHANGELOG.md
@@ -16,8 +16,8 @@ and this project adheres to
 
 ## Fixed
 
-- 🐛(backend) gitlab oicd userinfo endpoint #232
-
+- 🐛 (backend) gitlab oicd userinfo endpoint #232
+- ♻️ (backend) getting list of document versions available for a user #258
 
 ## [1.4.0] - 2024-09-17